+++ Server/StreetEmpire.Api/Program.cs
@@ -42,6 +42,7 @@ builder.Services.AddScoped<PlayerClock>();
 builder.Services.AddScoped<StandingsRecorder>();
 builder.Services.AddScoped<TerritoryService>();
 builder.Services.AddScoped<MarketService>();
+builder.Services.AddScoped<MuleService>();
 builder.Services.AddSingleton<StandingsSchedule>();
 builder.Services.AddScoped<PimpRoster>();
 builder.Services.AddScoped<AdminService>();